[MINOR] Separate in layers
